requirements.txt
anthropic
chromadb
sentence-transformers
pypdf
python-dotenv
numpy
python-docx
tiktoken
networkx
pyvis
